+++ Flow/Changes.md
@@ -2,8 +2,11 @@
 
 ## v2.5.1 (in development)
 ### New features
+- Added **LazilyUpdating** class, which performs lazy incremental updates regularly or when requested to do so
 - Added **UntilExternalFailureIterator**
 ### New methods
+- **IterableOnce**
+  - Added `.splitMapInto(...)`, `.splitFlatMapInto(...)` and `.splitInto(...)`
 - **ModelLike**
   - Added `.tryGet(...)`
 - **Seq** (**CollectionExtensions**)